URL
In this program, we first prompt the user to enter the original URL they want to shorten. We then use the pyshorteners library to shorten the URL using the TinyURL service, which is one of the many URL shortening services supported by the library. Finally, we print the shortened URL to the console.